Rename `OfferIssuerId`
The `OfferIssuerId` was previously named `OfferNodeId`, but its usage
changed when it was renamed: it doesn't necessarily match a lightning
`node_id` and can instead be a completely unrelated public key.

It can now be included in combination with blinded paths: when that
happens, the blinded paths must be used in priority to reach the node
as the `issuer_id` may not match a network `node_id` at all (or may
even match an unrelated node).