send tx_sigs even if commit_sig was already received
Due to a bug in lightning-kmp 1.5.x, we were sometimes not sending our
`tx_sigs` first, when we had to, causing both peers to be stuck waiting
for eachother's `tx_sigs`.

When receiving remote's `commit_sig`, we were also operating under the
assumption that if the local commit was already fully signed, then we
already had assessed whether or not we should send our local `tx_sigs`
first, because we don't persist channel data when receiving remote
`commit_sig`, but we do when we send our `tx_sigs`.

But other operations in the channel may cause the channel to be
persisted, causing our channel state to contain a fully signed local
commit without having sent our local `tx_sigs`. So, even after fixing
the `tx_sigs` ordering bug, we would never send it.