Cancel splice early when missing funds
When we are splicing out or doing a CPFP on an existing splice, we use
our channel balance to pay fees and the LSP doesn't contribute at all.
We must have enough funds to pay the mining fees, otherwise we will go
through all of the `interactive-tx` protocol but fail at the end when
our peer will notice that we're not paying the feerate we said we would.

We now fail early in that case, which also lets us provide a better
error explaining why the splice attempt was cancelled.