Rework liquidity purchase storage in DB (#749)
We previously used a dedicated `InboundLiquidityOutgoingPayment` entry
in our outgoing payments DB for every liquidity purchase: the issue is
that for new channels, the full information about mining fee was split
between this entry and the `NewChannelIncomingPayment` entry in the
incoming payments DB, which was confusing. This would become even more
confusing as we add support for purchasing liquidity in more scenarios.

We re-work this to ensure that:

- there is a single entry in either the incoming or outgoing payments DB
  for each on-chain transaction: on-chain payments may use the `txId` as
  part of their primary key or a DB index can be used
- if a liquidity purchase is contained in this on-chain transaction, the
  purchase details are directly included in the DB entry (no need for
  joining tables unless that's an explicit choice of the wallet devs)
- when we receive an incoming lightning payment that pays for a previous
  liquidity purchase, we directly store the purchase details with this
  incoming payment, which removes the need for an additional lookup
  and we update the on-chain operation to mark it as received
- we add an optional liquidity purchase to a few existing on-chain
  payments, even though they are currently unused and always `null`:
  we can easily add support for those scenarios in the future

Note that since the binary serialization for payments hasn't been
officially released yet, we break backwards-compatibility with the
previous encoding found on `master`.