Fix `LightningIncomingPayment.receiveAmount` (#751)
The `receiveAmount` must be the total contribution to the balance. Since we are now attaching the liquidity purchase to the `LightningIncomingPayment`, we must correctly account for fees paid from the balance and paid directly from HTLCs.

If a liquidity purchase was involved, then a fee will already have been paid upfront directly from our channel balance (even in the `LiquidityAds.PaymentDetails.FromFutureHtlc` case). We must deduct it from the amount received via `Part.Htlc`.