Fix feerate adjustment for on-the-fly liquidity
When purchasing on-the-fly liquidity without any channel balance, we
inflate the feerate because we cannot contribute to the mining fees,
so that the final feerate is close to the target feerate.

However, the ratios we were using weren't properly computed which
resulted in transactions below 1 sat/byte, which thus couldn't be
relayed. We now compute the real ratios depending on the number of
inputs/outputs the LSP will contribute, and adjust the feerate based
on those ratios.