Refactor watcher types
Watch types (and mostly the `BitcoinEvent` sealed class) were confusing
because we imported code from `eclair` and never adapted it to mobile
wallets where we have less complex watch scenarios.

We refactor this and rename the different kind of events to match what
we're using. This is a first step towards doing exhaustive pattern
matching when a watch triggers, which we'll do in the next commit.

We also add a few helper variables and move funding helpers.