Refactor watcher types (#760)
Watch types (and mostly the `BitcoinEvent` sealed class) were confusing
because we imported code from `eclair` and never adapted it to mobile
wallets where we have less complex watch scenarios.

We refactor this and rename the different kind of events to match what
we're using. We then use exhaustive pattern matching when a watch
triggers, which is particularly important in the `Closing` state where we
want to make sure we're not missing any scenario.