Store per-commitment `short_channel_id` (#740)
Each commitment has a different `short_channel_id`: it should thus be
stored inside the `Commitment` object. We keep a `shortChannelId` field
in the channel state because it contains:

- our scid alias if the channel is 0-conf
- otherwise the real scid of the initial funding transaction

This should be replaced by the local/remote aliases once we support
`scid_alias` instead of using our custom `PeerId`. Note that our current
`PeerId` trick is compatible with `scid_alias` since we send it using
the same TLV in `channel_ready`.