Bolt11 invoices: do not re-encode amounts
Bolt11 amounts may not be minimally encoded (the specs says they SHOULD be but it's not a hard requirement).
We should preserve the original amount encoding: if it was not minimally encoded and we re-encode it using its
minimal representation (for example "5" instead of "5000m") then read(write()) will fail to verify the invoice checksum
and return an "invoice isn't canonically encoded" error.