Send normal failures for Bolt12 payments
Since we are always using 1-hop blinded paths from our LSP to ourselves,
there are no public intermediate nodes that could be probed inside that
blinded path. The requirement to always return malformed failure is thus
unnecessary (and harmful for the payer who doesn't know why the payment
failed). We now return normal failures in that case, which should be
forwarded by our LSP to the upstream nodes.

This also fixes an issue where non-trampoline blinded payments were not
correctly failed because the extraction of the onion shared secret did
not uses the `path_key` to derive the blinded onion decryption key.