Refactor deterministic offer functions
This commit refactors and simplifies the deterministic offer mechanism.
We expose only two functions in `NodeParams`: `defaultOffer` and
`randomOffer`. The details of the deterministic offer creation have been
moved to `OfferManager` and we removed one of the `createBlindedOffer`
overloads.

We also remove support for custom and additional TLVs: we don't have any
use-case for them for this feature, so there's no reason to potentially
introduce complexity.

We slightly change the tweaking: it was partially using the offer merkle
tree but also combining the nonce outside of it, which was confusing. We
simply create a tagged hash that commits to each of the offer fields.