Emit `ChannelCloseOutgoingPayment` as soon as a mutual close tx is known (#783)
As opposed to when the mutual close tx is confirmed.

This is a return to the behavior we had _before_ simple close. This behavior changed because with simple close, we now stay in `Negotiating` until a mutual close tx is confirmed. Before, we would go to `Closing` state as soon as a mutual close tx was known.

Note that we are currently not handling the case where a concurrent mutual close tx confirms (they can be rbfed). But we were already not handling the case where a concurrent alternate close type confirms (e.g. a force close overrides a mutual close).