Keep features as a ByteVector in Bolt12 TLVs (#790)
We always encode features using as few bytes as needed, however other implementations may use more bytes than needed. When they do in Bolt12 TLVs, our reencoding of the TLV is different and changes the hash. Instead we should keep the original encoding and only decode the field when needed.