less aggressive security margin for otf payments
Liquidity fees depend on mining fees which are volatile, and may
increase between the time when a user accepts otf liquidity, and
the time when the LSP allocates the liquidity. We need to make
sure that the user is able to afford the increase of cost, so we
require the incoming payment + fee credit to be 2x the estimated
fee.

But that is overkill because only the mining fee is volatile, not
the liquidity fee. On top of that, the liquidity fee can be much
greater than the mining fee for some use cases (e.g. phoenixd).