Remove duplication of htlcs storage for `SpliceStatus` (#796)
When creating a splice, we were storing two full copies of htlcs (for local and remote commits), that were already present in the main commitments.

Removing this duplication makes it possible to increase the default `maxAcceptedHtlcs` to 12.

Best reviewed commit by commit, there are some preparatory commits, but the actual change is just one small commit.