Add a new peer storage format to handle legacy backups (#807)
Our current peer storage data format is `encrypt(list(channel))`.

We add a new format: `list(encrypt(channel))`.

This allows handling of legacy proprietary per-channel backups: on the LSP side, we just need to concatenate legacy backups and set that as the peer storage, and then reject channel updates with the legacy backup method (we actually just reject connections from peers that do not support the peer storage backup).

Note that we do not yet remove the "restore from `channel_reestablish`" logic, so that we don't have to synchronize the phoenix update with the LSP update.