Resign next remote commit on reconnection
We previously retransmitted our last `commit_sig` on reconnection if it
hadn't been received by our peer, without changing it. This is an issue
when a commitment has been deactivated while offline, because the batch
size will be incorrect. This can also be an issue for taproot channels
when remote nodes don't use deterministic nonce derivation, because
their nonce may be different on reconnection and our previous
`commit_sig` would thus not be valid anymore.