Send splice `commit_sig`s as a batch (#787)
We introduce a `CommitSigBatch` class to group `commit_sig` messages
when splice transactions are pending. We use this class to ensure that
all the `commit_sig` messages in the batch are sent together to our
peer, without any other messages in-between.

When we receive `commit_sig` messages that contain the `batch` TLV, we
group them directly in the `PeerConnection` before relaying them as a
batch to the channel.