Improve splice preparation performance (#829)
We previously fetched the number of confirmations of previous funding
transactions (via electrum or mempool.space) whenever estimating fees
for a splice. This is dumb, we only need to fetch that for funding txs
that are in the `LocalFundingStatus.UnconfirmedTx` status, which avoids
making an API call in most cases.