Update `_channels` map before processing actions
We aren't consistent with regards to updating the `_channels` map and
calling `processActions`: sometimes we do it first, sometimes second.

It's really hard to reason about this and figure out whether this may
have side effects or not, but it seems more consistent to update the
channel state first, then process its follow-up actions.