Improve logs during reconnection (#835)
Include commitments MDC in `Offline` and `Syncing`: this helps figure
out reconnection issues where the splice state doesn't match.
We also include more logs in a few related places.

We aren't consistent with regards to updating the `_channels` map and
calling `processActions`: sometimes we do it first, sometimes second.

It's really hard to reason about this and figure out whether this may
have side effects or not, but it seems more consistent to update the
channel state first, then process its follow-up actions.